Create Exercise 1 : DVD Rentals
Get a list of all rentals which are out (have not been returned). How do we identify these films in the database?

Get a list of all customers who have not returned their rentals. Make sure to group your results.

Get a list of all the Action films with Joe Swank.
Before you start, could there be a shortcut to getting this information? Maybe a view?